gradleビルドに変更した(APKの動作はICSのみ確認)
-Xlintを有効にしてビルドすればわかるが、最近のAndroidでは非推奨なAPIを大量に含んでいるため
最近のAndroidでは動かない可能性が高い
手持ちのAndroid ICS実機でDebugビルドのAPKが動くのは確認した

+++ IDPWMemoICS/src/main/java/neetsdkasu/idpwmemoics/MemoMenuDialogFragment.java
@@ -35,20 +35,17 @@ public class MemoMenuDialogFragment extends DialogFragment
     public void onClick (View v) {
         Listener listener = (Listener) getActivity();
         String memoName = getArguments().getString(MEMO_NAME);
-        switch (v.getId()) {
-            case R.id.memo_menu_dialog_open:
+        final int id = v.getId();
+
+        if (id == R.id.memo_menu_dialog_open)
                 listener.openMemo(memoName);
-                break;
-            case R.id.memo_menu_dialog_export:
+        else if (id == R.id.memo_menu_dialog_export)
                 listener.exportMemo(memoName);
-                break;
-            case R.id.memo_menu_dialog_change_password:
+        else if (id == R.id.memo_menu_dialog_change_password)
                 listener.changeMemoPassword(memoName);
-                break;
-            case R.id.memo_menu_dialog_delete:
+        else if (id == R.id.memo_menu_dialog_delete)
                 listener.deleteMemo(memoName);
-                break;
-        }
+
         dismiss();
     }
